add report helper class
